[chore] Minor fixes, code guidelines
