+++ src/parsing/compileAstToExpression.ts
@@ -16,7 +16,7 @@ import PrecedingAxis from '../expressions/axes/PrecedingAxis';
 import PrecedingSiblingAxis from '../expressions/axes/PrecedingSiblingAxis';
 import SelfAxis from '../expressions/axes/SelfAxis';
 import IfExpression from '../expressions/conditional/IfExpression';
-import StackTraceGenerator from '../expressions/debug/StackTraceGenerator';
+import StackTraceGenerator, { SourceRange } from '../expressions/debug/StackTraceGenerator';
 import ForExpression from '../expressions/ForExpression';
 import FunctionCall from '../expressions/functions/FunctionCall';
 import InlineFunction from '../expressions/functions/InlineFunction';
@@ -255,12 +255,13 @@ function compileTest(ast: IAST, compilationOptions: CompilationOptions): TestAbs
 	}
 }
 
-function stackTrace(ast, compilationOptions) {
-	const [, location, innerExpression] = ast;
+function stackTrace(ast: IAST, compilationOptions: CompilationOptions) {
+	const location = ast[1] as SourceRange;
+	const innerExpression = ast[2] as IAST;
 
-	let nextCompilableExpression = innerExpression;
+	let nextCompilableExpression: IAST = innerExpression;
 	while (nextCompilableExpression[0] === 'x:stackTrace') {
-		nextCompilableExpression = nextCompilableExpression[2];
+		nextCompilableExpression = nextCompilableExpression[2] as IAST;
 	}
 
 	return new StackTraceGenerator(
